Update the project to Xcode 9 and Swift 4

+++ SwiftRadioUITests/SwiftRadioUITests.swift
@@ -44,10 +44,10 @@ class SwiftRadioUITests: XCTestCase {
     
     func assertStationsPresent() {
         let numStations:UInt = 4
-        XCTAssertEqual(stations.count, numStations)
+        XCTAssertEqual(stations.count, Int(numStations))
         
         let texts = stations.staticTexts.count
-        XCTAssertEqual(texts, numStations * 2)
+        XCTAssertEqual(texts, Int(numStations * 2))
     }
     
     func assertHamburgerContent() {
